+++ generator/src/main/kotlin/com/kylemayes/generator/registry/Filter.kt
@@ -28,21 +28,7 @@ fun Registry.filterEntities(): Registry {
 // ===============================================
 
 /** Vulkan extensions that are specifically not supported by `vulkanalia`. */
-private val unsupportedExtensions = setOf(
-    // Video extensions depend on a bunch of types not defined in the Vulkan
-    // API registry but instead live in some additional Vulkan headers. Maybe
-    // support could be added using `bindgen` to generate Rust types from these
-    // additional Vulkan headers?
-    "VK_EXT_video_decode_h264",
-    "VK_EXT_video_decode_h265",
-    "VK_EXT_video_encode_h264",
-    "VK_EXT_video_encode_h265",
-    "VK_KHR_video_decode_h264",
-    "VK_KHR_video_decode_h265",
-    "VK_KHR_video_decode_queue",
-    "VK_KHR_video_encode_queue",
-    "VK_KHR_video_queue",
-)
+private val unsupportedExtensions = emptySet<String>()
 
 /** Gets whether a Vulkan extension is supported by `vulkanalia`. */
 private fun Extension.isSupported() =